Created a really long README.md
Just an explanation on why this thing is so shit. The entire reason I posted this on here even though it's an abandoned project is because I want other people to learn from my mistakes, I hope I formulated all the mistakes I made well enough to make that possible.